+++ src/view/exceptions.py
@@ -19,6 +19,7 @@ __all__ = (
     "ConfigurationError",
     "NeedsDependencyError",
     "InvalidTemplateError",
+    "TypeValidationError",
 )
 
 
@@ -106,3 +107,6 @@ class NeedsDependencyError(ViewError):
 
 class InvalidTemplateError(ViewError):
     """Something is wrong with a template."""
+
+class TypeValidationError(TypeError, ViewError):
+    """Could not assign the object to the target type."""